moved isPlatform util to common/platform
