+++ src/services/appUpdater.ts
@@ -12,7 +12,7 @@ import {
 import fetch from 'node-fetch';
 import { logErrorSentry } from './sentry';
 import ElectronLog from 'electron-log';
-import { isPlatform } from '../utils/main';
+import { isPlatform } from '../utils/common/platform';
 
 const FIVE_MIN_IN_MICROSECOND = 5 * 60 * 1000;
 const ONE_DAY_IN_MICROSECOND = 1 * 24 * 60 * 60 * 1000;
